runners: Add option to specify Ubuntu nightly kernel
Allow installing a custom kernel from the Ubuntu nightly packages
page to test ZFS against new kernels.  Specifically, add a text
box to specify the URL you want the runner to scrape and install
from.  For example:

https://kernel.ubuntu.com/mainline/v6.14-rc7/

This option only works on workflow_dispatch, so the user will need to
enter in the URL on the actions page and run the workflow manually.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>