RPM: Hold back incompatible kernel packages on Fedora
A user reported that when your upgrade your kernel packages on Fedora
with ZFS installed, only the kernel-devel package gets held back to the
ZFS-supported version, but not the other kernel packages. So if ZFS only
supports the 6.13 kernel, Fedora will still happily upgrade the kernel
RPMs to 6.14, but hold back kernel-devel at 6.13, for example.

This commit includes version checks for the 'kernel', 'kernel-core' and
'kernel-modules-core' packages.

Closes: #17265

Original-patch-by: @jkool702
Signed-off-by: Tony Hutter <hutter2@llnl.gov>