ZTS: Use random numbers for 'fill_fs'
A number of tests use 'fill_fs' to generate junk data.  The problem
is that many of those tests are not passing the 'R' option to
generate *random* data.  That results in unnecessary compression and
potential test timeouts.

Convert those tests to use 'R'.

Also add a comment to file_write.c to explain the psudorandom XOR
method it uses.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>